Mostly styles and aesthetics changes
Added UI feedback for if URL RegEx test is invalid.  Added text-shadow
to H1 elements.  Added .ng-not-empty class for inputs so that empty
inputs do not incorrectly show as “valid”.  Added less variables so
that IMGs in the placeholder containers scale proportionally (to a
certain limit).  Added a PUG for the Home view, and added placeholder
text for some text fields in the “add new” view.